requirements.txt
fdb
fastapi
uvicorn
streamlit
pandas
plotly
python-dotenv
requests
openpyxl
sshtunnel
paramiko==2.12.0
